fix: remove the assertion
It is not clear to me when this happens, but from how
the sync function is written there any reason to assert and fails.

Also, while start a new negotiation with a new peer, looks like pretty
common get a peer with a different height that maybe is the wrong one?

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>