feat: get an utxo from the mempool
There are usage also with the lite clients where it is needed
get utxo by an `txid` and an `idx`.

this is an initial approach where we look inside the mempool
by a txid and then we look inside the transactions output, but
this leave open a problem that is what we do when the transaction
is not inside the mempool? I think we need another approach where
from the network we need to fetch just a utxo.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>