feat: add the possibility to query the fee estimation by block
We should allow get an estimation fee by block and keep
maybe a mix size of the BTree where we said that > 100 block the fee
is equal to the 100, but also we could use a smart way to store the
bee by range, like 0-10 blocks and so on.

However, for now it is just a monkey implementation, because I would
like to implement a Estimator interface and expose this to the ecosystem
to allow more smarter way to estimate the fee.

All this to justify my current monkey implementation :)

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>