fix: Fix crash caused by block duplication during loading
Previously, the code would crash with the assertion error shown below:

```
2023-04-26T11:51:18.340Z INFO client Loading block headers from store..
thread '<unnamed>' panicked at 'assertion failed: `(left == right)`
  left: `2431071`,
 right: `2431070`', /home/vincent/.cargo/git/checkouts/nakamoto-90f2d5c288bae2ec/8744d43/chain/src/block/cache.rs:169:9
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: Disconnected', lampo-nakamoto/src/lib.rs:85:42
```

After investigating the issue, I found that there was a duplication in the block
header that was causing the problem, but I had no idea why this problem
is happening, so idea could be chain reorg.

The problem with this duplication is that it leaves the storage in the wrong
state, as there can be a mismatch between the blocks and headers. This is
because the headers are stored in a hash map, and the duplication replaces
the previous item.

To solve this issue, I added a check to detect the duplication and
truncate the storage and chain by keeping track of the safe height.

Some open question are if we need to allow to insert block duplicated in
the store at all? and if no how we manage this case?

Currently the best solution that I found is check if this happens during
the reload, and fix it.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>