[p2p] relax the contraints on the range
Some users report a bug to us with the following assertion

thread '<unnamed>' panicked at ..../.cargo/git/checkouts/nakamoto-75926082020ea03a/08203fd/p2p/src/fsm/cbfmgr.rs:449:9:
assertion failed: *range.end() <= self.filters.height()
flutter: Failed to join thread

The problem here may be that when nakamoto sync a new block
(the header storage is not in sync) and when we are going to
call `getcfiltes` we will have a crash.

I am still not sure that this is the right solution
because I do not remember how all the BIP 157 workflow looks
like, so I should double check this before open a patch.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>